feat: remove order validationContract
This feature is not desired for MVP, so we remove it in this commit

Also cleans up ordervalidator a bit, moving to an inherited contract
with local mapping instead of library with admittedly weird calling
semantics

+++ test/executors/DirectTakerExecutor.t.sol
@@ -5,8 +5,8 @@ import {Test} from "forge-std/Test.sol";
 import {DirectTakerExecutor} from "../../src/sample-executors/DirectTakerExecutor.sol";
 import {DutchLimitOrderReactor, DutchLimitOrder} from "../../src/reactor/dutch-limit/DutchLimitOrderReactor.sol";
 import {DutchLimitOrderExecution} from "../../src/reactor/dutch-limit/DutchLimitOrderStructs.sol";
-import {MockERC20} from "../../src/test/MockERC20.sol";
-import {Output, TokenAmount, OrderInfo, ResolvedOrder} from "../../src/interfaces/ReactorStructs.sol";
+import {MockERC20} from "../util/mock/MockERC20.sol";
+import {Output, TokenAmount, OrderInfo, ResolvedOrder} from "../../src/lib/ReactorStructs.sol";
 import {PermitPost, Permit} from "permitpost/PermitPost.sol";
 import {OrderInfoBuilder} from "../util/OrderInfoBuilder.sol";
 import {OutputsBuilder} from "../util/OutputsBuilder.sol";
@@ -51,18 +51,15 @@ contract DirectTakerExecutorTest is Test, PermitSignature {
     }
 
     function testReactorCallback() public {
-        uint inputAmount = ONE;
-        uint outputAmount = ONE;
+        uint256 inputAmount = ONE;
+        uint256 outputAmount = ONE;
 
         Output[] memory outputs = new Output[](1);
         outputs[0].token = address(tokenOut);
         outputs[0].amount = outputAmount;
         ResolvedOrder[] memory resolvedOrders = new ResolvedOrder[](1);
-        ResolvedOrder memory resolvedOrder = ResolvedOrder(
-            OrderInfoBuilder.init(address(dloReactor)),
-            TokenAmount(address(tokenIn), inputAmount),
-            outputs
-        );
+        ResolvedOrder memory resolvedOrder =
+            ResolvedOrder(OrderInfoBuilder.init(address(dloReactor)), TokenAmount(address(tokenIn), inputAmount), outputs);
         resolvedOrders[0] = resolvedOrder;
         bytes memory fillData = abi.encode(taker, dloReactor);
         tokenIn.mint(address(directTakerExecutor), inputAmount);
@@ -73,7 +70,7 @@ contract DirectTakerExecutorTest is Test, PermitSignature {
     }
 
     function testReactorCallback2Outputs() public {
-        uint inputAmount = ONE;
+        uint256 inputAmount = ONE;
 
         Output[] memory outputs = new Output[](2);
         outputs[0].token = address(tokenOut);
@@ -81,11 +78,8 @@ contract DirectTakerExecutorTest is Test, PermitSignature {
         outputs[1].token = address(tokenOut);
         outputs[1].amount = ONE * 2;
         ResolvedOrder[] memory resolvedOrders = new ResolvedOrder[](1);
-        ResolvedOrder memory resolvedOrder = ResolvedOrder(
-            OrderInfoBuilder.init(address(dloReactor)),
-            TokenAmount(address(tokenIn), inputAmount),
-            outputs
-        );
+        ResolvedOrder memory resolvedOrder =
+            ResolvedOrder(OrderInfoBuilder.init(address(dloReactor)), TokenAmount(address(tokenIn), inputAmount), outputs);
         resolvedOrders[0] = resolvedOrder;
         bytes memory fillData = abi.encode(taker, dloReactor);
         tokenOut.mint(taker, ONE * 3);
@@ -115,12 +109,7 @@ contract DirectTakerExecutorTest is Test, PermitSignature {
                 vm,
                 makerPrivateKey,
                 address(permitPost),
-                Permit({
-                    token: address(tokenIn),
-                    spender: address(dloReactor),
-                    maxAmount: ONE,
-                    deadline: order.info.deadline
-                }),
+                Permit({token: address(tokenIn), spender: address(dloReactor), maxAmount: ONE, deadline: order.info.deadline}),
                 0,
                 uint256(orderHash)
             ),
